Refactor header/footer for scss (needs clean up more)
