+++ frontend/src/App.vue
@@ -1,8 +1,8 @@
 <script lang="ts">
 
   import { RouterView } from 'vue-router'
-  import Header from '@/components/Header.vue'
-  import Footer from '@/components/Footer.vue'
+  import Header from '@/components/header/Header.vue'
+  import Footer from '@/components/footer/Footer.vue'
 
   export default {
     components: {
